feat: refactor network map
