+++ wasm-app/src/components/modules/networkgraph/color.ts
@@ -1,0 +1,19 @@
+export const color = {
+  white: "#fff",
+  stroke: {
+    default: "#4d4d4f",
+    active: "rgba(83, 139, 255, 1)",
+  },
+  fill: {
+    default: "#2d323d",
+    active: "#2d4169",
+    interaction100: "rgba(83, 139, 255, 0.1)",
+    interaction200: "rgba(83, 139, 255, 0.2)",
+  },
+
+  active: "#125AED",
+  idle: "#FFA800",
+  idleBackground: "#302d27",
+  error: "#EB4136",
+  errorBackground: "#2e222c",
+};
