--- wasm-app/src/components/modules/networkgraph/data.ts
@@ -1,117 +1,0 @@
-export const workloads = [
-  {
-    uuid: "7431bb4f-cae8-4dbe-a542-d6f52c893271",
-    workloadName: "default-workload-1",
-    kind: "deployment",
-    size: 60,
-    from: [
-      {
-        workloadId: "7f2552b4-ab40-4120-a6d9-16507024922b",
-        status: 1,
-      },
-      {
-        workloadId: "1b8892b1-58bc-464f-9401-b31eb2a9db99",
-        status: 4,
-      },
-    ],
-    to: [],
-  },
-  {
-    uuid: "7f2552b4-ab40-4120-a6d9-16507024922b",
-    workloadName: "default-ad-service",
-    kind: "deployment",
-    size: 40,
-    from: [],
-    to: [
-      {
-        workloadId: "7431bb4f-cae8-4dbe-a542-d6f52c893271",
-        status: 2,
-      },
-    ],
-  },
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db99",
-    workloadName: "default-ad-service",
-    kind: "deployment",
-    size: 74,
-    from: [
-      {
-        workloadId: "7f2552b4-ab40-4120-a6d9-16507024922b",
-        status: 3,
-      },
-    ],
-    to: [],
-  },
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db99xx",
-    workloadName: "default-ad-service",
-    kind: "deployment",
-    from: [
-      {
-        workloadId: "7f2552b4-ab40-4120-a6d9-16507024922b",
-        status: 4,
-      },
-    ],
-    to: [],
-  },
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db94",
-    workloadName: "axxxx",
-    kind: "deployment",
-    from: [
-      {
-        workloadId: "7f2552b4-ab40-4120-a6d9-16507024922b",
-        status: 4,
-      },
-    ],
-    to: [],
-  },
-
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db93",
-    workloadName: "workload name 1",
-    kind: "deployment",
-    from: [],
-    to: [],
-  },
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db92",
-    workloadName: "workload name 2",
-    kind: "deployment",
-    size: 40,
-    from: [
-      {
-        workloadId: "1b8892b1-58bc-464f-9401-b31eb2a9db93",
-        status: 4,
-      },
-    ],
-    to: [],
-  },
-
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db93aa",
-    workloadName: "holly ***",
-    kind: "deployment",
-    from: [],
-    to: [],
-  },
-  {
-    uuid: "1b8892b1-58bc-464f-9401-b31eb2a9db92bb",
-    workloadName: "holly ***",
-    kind: "deployment",
-    size: 40,
-    from: [
-      {
-        workloadId: "1b8892b1-58bc-464f-9401-b31eb2a9db93aa",
-        status: 4,
-      },
-    ],
-    to: [],
-  },
-];
-
-export const addIconSvg = `<svg width="10" height="12" viewBox="0 0 10 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M5.36716 0.861942L9.03382 2.23694C9.23356 2.31181 9.40567 2.44587 9.52717 2.62119C9.64867 2.7965 9.71375 3.00473 9.71373 3.21804V6.02723C9.71375 6.90274 9.46996 7.76095 9.00969 8.50571C8.54942 9.25047 7.89085 9.85235 7.10778 10.2439L5.35092 11.1218C5.24179 11.1764 5.12146 11.2048 4.99944 11.2048C4.87743 11.2048 4.75709 11.1764 4.64797 11.1218L2.89111 10.2434C2.10803 9.85182 1.44946 9.24995 0.989193 8.50519C0.528922 7.76043 0.285138 6.90221 0.285156 6.0267V3.21856C0.285023 3.00517 0.350059 2.79682 0.471563 2.6214C0.593068 2.44598 0.765245 2.31184 0.965061 2.23694L4.63173 0.861942C4.86882 0.77307 5.13007 0.77307 5.36716 0.861942ZM4.99944 3.90266C4.87114 3.90267 4.74731 3.94978 4.65144 4.03503C4.55556 4.12029 4.49431 4.23776 4.4793 4.36518L4.47563 4.42647V5.47409H3.42801C3.29451 5.47423 3.16609 5.52536 3.06901 5.61701C2.97193 5.70866 2.91351 5.83392 2.90569 5.9672C2.89786 6.10048 2.94122 6.23171 3.02691 6.33409C3.1126 6.43647 3.23415 6.50227 3.36673 6.51804L3.42801 6.5217H4.47563V7.56932C4.47578 7.70283 4.5269 7.83125 4.61855 7.92833C4.7102 8.02541 4.83547 8.08383 4.96874 8.09165C5.10202 8.09948 5.23326 8.05611 5.33564 7.97042C5.43802 7.88473 5.50382 7.76318 5.51959 7.63061L5.52325 7.56932V6.5217H6.57087C6.70438 6.52156 6.83279 6.47043 6.92987 6.37878C7.02695 6.28713 7.08538 6.16187 7.0932 6.02859C7.10102 5.89531 7.05766 5.76408 6.97197 5.6617C6.88628 5.55932 6.76473 5.49352 6.63216 5.47775L6.57087 5.47409H5.52325V4.42647C5.52325 4.28754 5.46806 4.15431 5.36983 4.05608C5.2716 3.95784 5.13837 3.90266 4.99944 3.90266Z" fill="white"/>
-<path d="M5.36716 0.861942L9.03382 2.23694C9.23356 2.31181 9.40567 2.44587 9.52717 2.62119C9.64867 2.7965 9.71375 3.00473 9.71373 3.21804V6.02723C9.71375 6.90274 9.46996 7.76095 9.00969 8.50571C8.54942 9.25047 7.89085 9.85235 7.10778 10.2439L5.35092 11.1218C5.24179 11.1764 5.12146 11.2048 4.99944 11.2048C4.87743 11.2048 4.75709 11.1764 4.64797 11.1218L2.89111 10.2434C2.10803 9.85182 1.44946 9.24995 0.989193 8.50519C0.528922 7.76043 0.285138 6.90221 0.285156 6.0267V3.21856C0.285023 3.00517 0.350059 2.79682 0.471563 2.6214C0.593068 2.44598 0.765245 2.31184 0.965061 2.23694L4.63173 0.861942C4.86882 0.77307 5.13007 0.77307 5.36716 0.861942ZM4.99944 3.90266C4.87114 3.90267 4.74731 3.94978 4.65144 4.03503C4.55556 4.12029 4.49431 4.23776 4.4793 4.36518L4.47563 4.42647V5.47409H3.42801C3.29451 5.47423 3.16609 5.52536 3.06901 5.61701C2.97193 5.70866 2.91351 5.83392 2.90569 5.9672C2.89786 6.10048 2.94122 6.23171 3.02691 6.33409C3.1126 6.43647 3.23415 6.50227 3.36673 6.51804L3.42801 6.5217H4.47563V7.56932C4.47578 7.70283 4.5269 7.83125 4.61855 7.92833C4.7102 8.02541 4.83547 8.08383 4.96874 8.09165C5.10202 8.09948 5.23326 8.05611 5.33564 7.97042C5.43802 7.88473 5.50382 7.76318 5.51959 7.63061L5.52325 7.56932V6.5217H6.57087C6.70438 6.52156 6.83279 6.47043 6.92987 6.37878C7.02695 6.28713 7.08538 6.16187 7.0932 6.02859C7.10102 5.89531 7.05766 5.76408 6.97197 5.6617C6.88628 5.55932 6.76473 5.49352 6.63216 5.47775L6.57087 5.47409H5.52325V4.42647C5.52325 4.28754 5.46806 4.15431 5.36983 4.05608C5.2716 3.95784 5.13837 3.90266 4.99944 3.90266Z" fill="white"/>
-</svg>
-`;
