+++ wasm-app/src/svg.d.ts
@@ -2,3 +2,8 @@ declare module "*.svg" {
   const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
   export default content;
 }
+
+declare module "*.svg?inline" {
+  const content: string;
+  export default content;
+}
